renamed files; use the latest API
